% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_templates.R
\name{apply_cell_template}
\alias{apply_cell_template}
\title{Apply Cell Formatting and Data Template to Excel Sheet}
\usage{
apply_cell_template(
  wb,
  df,
  cell.template,
  column.names = FALSE,
  column.name,
  sheet.number,
  column.number,
  column.type,
  data.length,
  data.start.row,
  validation.type,
  validation.operator,
  validation.value.1,
  validation.value.2,
  formula.text,
  fontName,
  fontSize,
  fontColour,
  numFmt,
  border,
  borderColour,
  borderStyle,
  bgFill,
  fgFill,
  halign,
  valign,
  textDecoration,
  wrapText,
  textRotation,
  indent,
  locked,
  hidden
)
}
\arguments{
\item{wb}{A workbook object from openxlsx.}

\item{df}{A data frame containing the data to be written.}

\item{cell.template}{A data frame containing cell template information.}

\item{column.names}{Logical. Should column names be included?}

\item{column.name}{Name of the column in the data frame.}

\item{sheet.number}{Numeric or character identifier for the sheet.}

\item{column.number}{The column number to apply the template to.}

\item{column.type}{Type of the column ("data", etc.)}

\item{data.length}{Number of rows to apply the data to.}

\item{data.start.row}{The starting row for the data.}

\item{validation.type}{Type of validation ("date", "numeric", etc.)}

\item{validation.operator}{Operator for validation ("greaterThan", etc.)}

\item{validation.value.1}{First value for validation.}

\item{validation.value.2}{Second value for validation.}

\item{formula.text}{Text representation of the Excel formula.}

\item{fontName}{Font name for cell style.}

\item{fontSize}{Font size for cell style.}

\item{fontColour}{Font color for cell style.}

\item{numFmt}{Number format for cell style.}

\item{border}{Type of cell border.}

\item{borderColour}{Color of cell border.}

\item{borderStyle}{Style of cell border.}

\item{bgFill}{Background fill for cell style.}

\item{fgFill}{Foreground fill for cell style.}

\item{halign}{Horizontal alignment for cell style.}

\item{valign}{Vertical alignment for cell style.}

\item{textDecoration}{Text decoration for cell style.}

\item{wrapText}{Text wrap setting for cell style.}

\item{textRotation}{Text rotation angle for cell style.}

\item{indent}{Indentation for cell style.}

\item{locked}{Logical. Is the cell locked?}

\item{hidden}{Logical. Is the cell hidden?}
}
\value{
NULL. The function modifies the workbook object in place.
}
\description{
This internal function applies a range of styles, validation rules, formulas,
and data to specific cells in an Excel workbook using the openxlsx package.
}
\keyword{internal}
