% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{generate_xls_formula}
\alias{generate_xls_formula}
\title{Generate Excel Formulas}
\usage{
generate_xls_formula(text, rows, sheet.number, cell.template)
}
\arguments{
\item{text}{Text representation of the Excel formula.}

\item{rows}{Row range to use in the formula.}

\item{sheet.number}{Numeric or character identifier for the sheet.}

\item{cell.template}{A data frame containing cell template information.}
}
\value{
A string representing the Excel formula.
}
\description{
This internal function generates Excel formulas based on a given text template
and additional parameters.
}
\keyword{internal}
