% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_example.R
\name{get_workbook_template}
\alias{get_workbook_template}
\title{Download Example File to Working Directory}
\usage{
get_workbook_template()
}
\description{
Copies a sample workbook template 'workbook_template.xlsx' from the
package's `extdata` folder to the user's current working directory.
}
\details{
The `get_workbook_template` function provides users with easy access to a
sample workbook template packaged with `excelr8`. This file can
be used as a basis for the user's work.


The example files are stored in the `extdata` directory of the
`excelr8` package. This function uses the `fs` package to handle
file operations.
}
\examples{
\dontrun{
# To copy the example files to your working directory
get_workbook_template()

# Check if the files are available in the working directory
list.files(pattern = "xlsx")
}

}
