% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_data}
\alias{write_data}
\title{Write Data to an Excel Sheet}
\usage{
write_data(
  wb,
  df,
  column.names,
  column.name,
  sheet.number,
  column.number,
  data.start.row,
  data.length
)
}
\arguments{
\item{wb}{A workbook object from openxlsx.}

\item{df}{Data frame to write.}

\item{column.names}{Boolean, should column names be written.}

\item{column.name}{Name of the column to select from the data frame.}

\item{sheet.number}{Numeric or character identifier for the sheet.}

\item{column.number}{Starting column number for the data.}

\item{data.start.row}{Starting row number for the data.}

\item{data.length}{Number of rows to write.}
}
\value{
A modified workbook object from openxlsx.
}
\description{
This internal function writes a data frame to a specific sheet and
position within an Excel workbook using the openxlsx package.
}
\keyword{internal}
