% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{xls_col}
\alias{xls_col}
\title{Convert Column Number to Excel Column Letter}
\usage{
xls_col(col_number)
}
\arguments{
\item{col_number}{Numeric index of the column.}
}
\value{
A string representing the Excel column letter.
}
\description{
This internal function converts a numeric column index to its Excel
column letter equivalent.
}
\keyword{internal}
