% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelr8.R
\name{excelr8}
\alias{excelr8}
\title{Write Excel File from Datasheet Template}
\usage{
excelr8(
  df,
  data.template.path,
  formatting.template.path = NULL,
  wb = NULL,
  output.file.name = NULL,
  unique_identifier = NULL,
  column.names = FALSE,
  messages_first_data_row = NULL
)
}
\arguments{
\item{df}{Data frame containing the data to be written to the Excel file.}

\item{data.template.path}{Character string specifying the path to the Excel file serving as the data template.}

\item{formatting.template.path}{Character string specifying the path to the Excel file serving as the formatting template. Default is NULL.}

\item{wb}{Workbook object from openxlsx. Default is NULL.}

\item{output.file.name}{Character string specifying the name of the output Excel file. Default is NULL.}

\item{unique_identifier}{This is necessary for use of the comments function.
Provide a unique identifier associated with the data to apply automatic comments to.
Default is NULL. If Null, the comments functionality is disabled.}

\item{column.names}{Logical indicating whether to include column names. Default is FALSE.}

\item{messages_first_data_row}{An intiger indicating which rows the messages start at. Default is NULL.
If NULL, the program looks to the Cell Info tab of the workbook template to infer a logical value.}
}
\value{
If output.file.name is NULL, returns the modified workbook object. Otherwise, it saves the workbook and returns NULL.
}
\description{
This function generates an Excel file based on data and optional formatting templates.
It allows you to apply cell, sheet, and workbook templates to an existing or new workbook.
}
\examples{
\dontrun{
excelr8(df = data.frame(a = 1:3, b = 4:6),
                              data.template.path = "path/to/data_template.xlsx",
                              output.file.name = "output.xlsx")
}

}
